requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
pandas-ta-classic
pyyaml
pydantic
aiohttp
python-dateutil
pytz
pyarrow
beautifulsoup4
lxml
requests
